move functions from main to a new module
